add nodejs backend for both conferences and video walls
